Split out `SettingsToolboxGroup` from `PlayerSettingsGroup`
